feat: Added a Monte Carlo Tree Search Agent
It assumes certain basic functions related to the STATE that is passed through, that need to be incorporated on that class. Also, an extra put_action() function has to be passed along in order to insert the best option encountered in the time given. See the document for more information.