fix: "Tree_Node_Class" - "backup_n_and_q" function
The update of n and q was inverse to the one expected on Monte Carlo Tree Search. This version uses the regular MAX approach (maximizes the solutions), and therefore the update had to be inverted back to normal. "n" gets incremented by one to force opening non-explored routes, and "b" gets incremented with the cost value for each new simulation made